app/requirements.txt
gunicorn==19.4.5
Flask==0.10.1
flask-restful==0.3.5
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
itsdangerous==0.24
wsgiref==0.1.2
celery==3.1.23
requests==2.9.1
python-magic==0.4.11
python-swiftclient==2.7.0